Mise à jour de la classe virtuelle CameraVirtuelle

+++ wxOpenCVMain/CameraAndor.h
@@ -211,7 +211,7 @@ virtual	char IsEMCCD(){return false;};
 virtual void DefEMCCDGain(int);
 virtual void DefModeGain(int);
 virtual void LireIntervalGain(int *x,int *y);
-virtual float TempsExposition (){return tempsExposition;};
+virtual double TempsExposition (){return tempsExposition;};
 virtual float TempsPose(){return 0;};
 virtual float GainEMCCD(){return 0;};
 virtual int ModeGainEMCCD(){return emCCDGainMode;};
